Fix datatype

+++ binance_bot/strategy/strategy_action.py
@@ -1,9 +1,11 @@
 from typing import Literal
 
+import numpy as np
+
 
 class StrategyAction:
 
-    def __init__(self, side: Literal["SELL", "BUY"], pair: str, quantity: float):
+    def __init__(self, side: Literal["SELL", "BUY"], pair: str, quantity: np.double):
         self.side = side
         self.pair = pair
         self.quantity = quantity
